Move some of the model and test creation to own function in different file. Needed to re-use them for scheduling tests
