+++ components/test-form/ParamsForm.js
@@ -20,8 +20,7 @@ export default class ParamsForm extends React.Component {
   }
 
   componentDidUpdate (prevProps, prevState, snapshot) {
-    if (!_.isEqual(this.props.schema, prevProps.schema) ||
-        !_.isEqual(this.props.default_params, prevProps.default_params)) {
+    if (!_.isEqual(this.props.schema, prevProps.schema)) {
       // eslint-disable-next-line react/no-did-update-set-state
       this.setState(
         {
